requirement.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.0
async-generator==1.10
attrs==22.2.0
Authlib==1.2.0
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
exceptiongroup==1.1.0
Flask==2.2.2
flower==1.2.0
gevent==22.10.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
outcome==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
td-ameritrade-python-api==0.3.5
tda-api==1.6.0
tomli==2.0.1
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
zope.event==4.6
zope.interface==5.5.2
